Upload file and delete files functionality

+++ src/app/modules/category/category.service.ts
@@ -72,6 +72,13 @@ const updateCategory = async (id: string, payload: Partial<ICategory>) => {
 };
 
 const deleteCategory = async (id: string) => {
+      const existingCategory = await Category.findById(id);
+    if (!existingCategory) {
+        throw new Error("Category not found.");
+    }
+        if (existingCategory.thumbnail) {
+        await deleteImageFromCLoudinary(existingCategory.thumbnail)
+    }
     await Category.findByIdAndDelete(id);
     return null;
 };
